refactor: move tanstack generator to separate directory
