+++ packages/cli/src/generators/tanstack-react/ts-factory/getIndexFactory.ts
